requirements.txt
Authlib==1.4.1
blinker==1.9.0
buildozer==1.5.0
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
csrf==0.1b1
distlib==0.3.9
filelock==3.17.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.1
MarkupSafe==3.0.2
msgspec==0.19.0
pexpect==4.9.0
platformdirs==4.3.6
psycopg2==2.9.10
ptyprocess==0.7.0
pycparser==2.22
pyotp==2.9.0
python-docx==1.1.2
python-dotenv==1.0.1
python-env==1.0.0
requests==2.32.3
sh==2.2.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.29.1
Werkzeug==3.1.3
WTForms==3.2.1
